pyproject.toml
[project]
name = "rival-agent-onboard"
version = "0.1.0"
description = "Conversational onboarding guide for new Rival team members"
authors = [{ name = "Rival" }]
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "requests>=2.31.0",
    "google-auth>=2.40.0",
    "google-cloud-firestore>=2.21.0",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "langchain-core>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "pyyaml>=6.0",
    "rival-agent-shared>=0.1.0",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "app.handlers"]
